# Fast GCD Inverse for Goldilocks 