Similar to #920 and #921 we can also speed up inversion in the Goldilocks field but doing a GCD based strategy.

We get about a `2x` speed up with just this PR with an additional `10-20%` once we merge #923.

| Field | Old Inv Speed (ns) | New Inv Speed (ns) | New + #923 Inv Speed (ns) |
| --- | --- | --- | --- |
| Goldilocks | 575 | 283 | 256 |
| Goldilocks Deg 2 Ext | 636 | 318 | 260 |
| Goldilocks Deg 5 Ext | 911 | 605 | 534 |